Running also on windows

+++ utils.py
@@ -37,8 +37,8 @@ def plot_contour(ax, func, *args):
 
 def generate_grid(grid_size, resolution, scale=1, should_ravel=True):
     print("creating a grid", 2 * resolution / scale)
-    y = np.linspace(-grid_size, grid_size, 2 * resolution / scale)
-    x = np.linspace(-grid_size, grid_size, 2 * resolution / scale)
+    y = np.linspace(-grid_size, grid_size, int(2 * resolution / scale))
+    x = np.linspace(-grid_size, grid_size, int(2 * resolution / scale))
     x_matrix, y_matrix = np.meshgrid(x, y)
     if should_ravel:
         return x_matrix.ravel(), y_matrix.ravel()
